Heartbeat refactoring
