+++ FleetDaemon/FleetDaemon.csproj
@@ -49,10 +49,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Heartbeat.cs" />
+    <Compile Include="Heartbeat\Heartbeat.cs" />
+    <Compile Include="MessageDispatcher\IMessageDispatcher.cs" />
+    <Compile Include="MessageDispatcher\LocalMessageDispatcher.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="RemoteFileManager.cs" />
+    <Compile Include="Heartbeat\RemoteFileManager.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
